fix: address PR review feedback
- Handle nullable dataset_item in metric_fn_with_interpolation
- Extract GEvalField to separate file for better maintainability
- Add test for None dataset_item scenario
